backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
face-recognition==1.3.0
dlib==19.22.1
scikit-learn==0.24.2
tensorflow==2.6.0
requests==2.26.0
mysql-connector-python==8.0.27
SQLAlchemy==1.4.27 